third_party/blink/web_tests/platform/generic/fast/filesystem/workers/read-directory-sync-many-expected.txt

[Worker] DirectoryReaderSync.readEntries() test with many entries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/read-directory-sync-many.js
PASS [Worker] numFiles is numFilesExpected
PASS [Worker] numDirectories is numDirectoriesExpected
PASS successfullyParsed is true

TEST COMPLETE
